Refactor server pages so overview is within the server, refactor routing
