+++ src/params/firestoreID.js
@@ -1,0 +1,4 @@
+/** @type {import('@sveltejs/kit').ParamMatcher} */
+export function match (param) {
+	return param.length === 20
+}
