+++ src/routes/[class]/[roomID=firestoreID]/+page.js
@@ -2,6 +2,7 @@
 export function load ({ data, params }) {
   // this corresponds to the `data` object
   return {
-    classID: params.class
+    classID: params.class,
+    roomID: params.roomID
   }
 }
