+++ src/routes/[class]/overview/+page.svelte
@@ -1,0 +1,8 @@
+<div use:portal={'main-content'}>
+  This is the overview page where you can clearly get a good sense of the server activities and 
+  easily sign up to teach / or crowdfund more teachers
+</div>
+
+<script>
+  import { portal, lazyCallable } from '/src/helpers/actions.js'
+</script>
