+++ src/routes/[class]/question/+page.svelte
@@ -1,0 +1,10 @@
+<ClassServerAskQuestion {classID}/>
+
+<script>
+  import ClassServerAskQuestion from '$lib/ClassServerAskQuestion.svelte'
+
+  export let data;
+  
+  let { classID } = data
+  $: ({ classID } = data)
+</script>
